refactor: Moved code into modules
