+++ app/Main.hs
@@ -6,53 +6,21 @@
 module Main where
 
 import Brick
-import Brick.Widgets.Center (center)
-import Data.List.Split (chunksOf)
+import Brick.Widgets.Center
+import Data.List.Split
 import qualified Data.Map as Map
 import GHC.Arr
 import Graphics.Vty
+import Player
+import World
 
-type World = [Level]
-type Level = [Room]
 type Name = ()
-type Coordinate = (Int, Int)
-
-data Player = Player
-    { name :: String
-    , pos :: Coordinate
-    }
-
-data Monster = Zombie | Ghost
-
-data Room = Room
-    { cells :: Array Coordinate Cell
-    , monsters :: Map.Map Coordinate Monster
-    }
-    deriving (Show)
-
-data VerticalDirection = Upwards | Downwards
-
-data Cell = Floor | Wall | Door | Stair VerticalDirection
 
 data GameState = GameState
     { player :: Player
     , world :: World
     }
     deriving (Show)
-
-instance Show Cell where
-    show Floor = ".."
-    show Wall = "##"
-    show Door = "λ "
-    show (Stair Upwards) = "Λ "
-    show (Stair Downwards) = "V "
-
-instance Show Monster where
-    show Zombie = "🧟"
-    show Ghost = "👻"
-
-instance Show Player where
-    show _ = "😎"
 
 app :: App GameState () Name
 app =
@@ -66,13 +34,6 @@ app =
             _ -> return ()
         , appStartEvent = return ()
         , appAttrMap = \_ -> attrMap defAttr []
-        }
-
-emptyRoom :: Room
-emptyRoom =
-    Room
-        { cells = listArray ((0, 0), (9, 9)) (replicate 100 Floor)
-        , monsters = Map.fromList [((2, 2), Zombie), ((4, 5), Ghost)]
         }
 
 drawGame :: GameState -> [Widget Name]
