+++ lambdabyrinth.cabal
@@ -65,7 +65,7 @@ executable lambdabyrinth
     main-is:          Main.hs
 
     -- Modules included in this executable, other than Main.
-    -- other-modules:
+    other-modules: Player World Monsters
 
     -- LANGUAGE extensions used by modules in this package.
     -- other-extensions:
